# docs/swagger.yaml
basePath: /api
definitions:
...
  create_role.CreateRoleRequest:
    properties:
      color:
        maxLength: 100
        minLength: 1
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - color
    - name
    type: object
  dessassign_role.DessassignRoleFromUserRequest:
    properties:
      role_id:
        maxLength: 100
        minLength: 1
        type: string
      user_id:
        maxLength: 100
        minLength: 1
        type: string
      workspace_id:
        maxLength: 100
        minLength: 1
        type: string
    required:
    - role_id
    - user_id
    - workspace_id
    type: object
...
  update_role.UpdateRoleRequest:
    properties:
      color:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
      permissions:
        minimum: 0
        type: integer
    required:
    - name
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@supchat.io
    name: Support API SupChat
  description: API pour le service de messagerie SupChat
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SupChat API
  version: "1.0"
paths:
...
  /api/workspace-invite-link/{token}/join:
    post:
      consumes:
      - application/json
      description: Permet à l'utilisateur authentifié de rejoindre un espace de travail
        en utilisant un lien d'invitation
      parameters:
      - description: Token d'invitation
        in: path
        name: token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Utilisateur ajouté à l'espace de travail
          schema:
            type: string
        "400":
          description: Token d'invitation manquant
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Token d'invitation invalide ou expiré
          schema:
            additionalProperties:
              type: string
            type: object
        "409":
          description: L'utilisateur est déjà membre de cet espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de l'ajout de l'utilisateur à l'espace de travail
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Rejoindre via lien d'invitation
      tags:
      - workspace
      - invite
...
  /api/workspaces/{workspace_id}/roles:
    get:
      consumes:
      - application/json
      description: Récupère la liste de tous les rôles disponibles dans un espace
        de travail spécifié
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Liste des rôles
          schema:
            items:
              additionalProperties:
                type: string
              type: object
            type: array
        "400":
          description: ID de l'espace de travail manquant
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes ou utilisateur non membre de l'espace
            de travail
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la récupération des rôles
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Liste des rôles
      tags:
      - workspace
      - role
    post:
      consumes:
      - application/json
      description: Crée un nouveau rôle avec un nom et une couleur dans l'espace de
        travail spécifié
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: Données du rôle à créer
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/create_role.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rôle créé avec succès
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Requête invalide ou données manquantes
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes pour créer un rôle
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la création du rôle
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Création d'un rôle
      tags:
      - workspace
      - role
  /api/workspaces/{workspace_id}/roles/{role_id}:
    delete:
      consumes:
      - application/json
      description: Supprime un rôle existant dans l'espace de travail spécifié
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: ID du rôle à supprimer
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: Rôle supprimé avec succès
          schema:
            type: string
        "400":
          description: ID de rôle manquant
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes pour supprimer un rôle
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Rôle non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la suppression du rôle
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Suppression d'un rôle
      tags:
      - workspace
      - role
    get:
      consumes:
      - application/json
      description: Récupère les informations détaillées d'un rôle spécifique dans
        un espace de travail
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: ID du rôle à récupérer
        in: path
        name: role_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Détails du rôle
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: ID du rôle manquant
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes pour voir les rôles
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Rôle non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la récupération du rôle
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Détails d'un rôle
      tags:
      - workspace
      - role
    put:
      consumes:
      - application/json
      description: Met à jour les informations d'un rôle existant (nom, couleur, permissions)
      parameters:
      - description: ID de l'espace de travail
        in: path
        name: workspace_id
        required: true
        type: string
      - description: ID du rôle à mettre à jour
        in: path
        name: role_id
        required: true
        type: string
      - description: Nouvelles informations du rôle
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/update_role.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rôle mis à jour avec succès
          schema:
            type: string
        "400":
          description: Requête invalide ou données incomplètes
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes pour gérer les rôles
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Rôle non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la mise à jour du rôle
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Mise à jour d'un rôle
      tags:
      - workspace
      - role
...
  /api/workspaces/{workspace_id}/roles/dessassign:
    post:
      consumes:
      - application/json
      description: Retire un rôle spécifique à un membre de l'espace de travail
      parameters:
      - description: Données de retrait du rôle
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dessassign_role.DessassignRoleFromUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Rôle retiré avec succès
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Requête invalide ou données incomplètes
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Permissions insuffisantes pour gérer les rôles
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Utilisateur ou rôle non trouvé
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors du retrait du rôle
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Retrait d'un rôle
      tags:
      - workspace
      - role
...
  /api/workspaces/discover:
    get:
      consumes:
      - application/json
      description: Récupère tous les espaces de travail publics disponibles pour rejoindre
      produces:
      - application/json
      responses:
        "200":
          description: Liste des espaces de travail publics
          schema:
            items:
              additionalProperties:
                type: string
              type: object
            type: array
        "401":
          description: Utilisateur non authentifié
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Erreur lors de la récupération des espaces de travail
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - ApiKeyAuth: []
      summary: Liste des espaces de travail publics
      tags:
      - workspace
      - discovery
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"